Updated generic delete confirmation modal, still needs more work

+++ app/components/main/confirmation/confirmation-modal-controller.js
@@ -2,8 +2,9 @@ angular.module('ehelseEditor').controller('ConfirmationModalController',['$rootS
 
     $scope.returnTrue = function() {
         $rootScope.confirmationValue = true;
-        $rootScope.deleteFieldById($rootScope.objectToDelete);
-        console.log($rootScope.confirmationValue);
+        if($rootScope.typeAsString == 'field'){
+            $rootScope.deleteFieldById($rootScope.objectToDelete);
+        }
     }
 
 }]);
